Run tests issue 16 (#24)
* Add option to keep virtual environment in Grader and VirtualEnvironment classes

* Add TestsCheck class to evaluate submitted code against test cases

* Add configuration for test checks in tests.json

* Add TestsCheck to checks factory for test evaluations

* Add --keep-venv option to CLI for retaining virtual environment after grading

* Add Pytest constants for cross-platform compatibility

* Add support for custom environment variables in command execution

* Update pytest version to 8.3.5 in grader requirements

* Add sample code to test the grader

* Enhance TestsCheck class with detailed docstrings for better clarity and maintainability

* Add unit tests for the TestsCheck class to verify scoring and logging behavior

* Rename the tests running check, as it was breaking VSCode test discovery

* Fix Python executable name for Windows to use 'python.exe'

* Add 'keep_venv' argument to get_args tests and normalize paths for Windows in structure validator tests

* Refactor RunTestsCheck to use process module for running commands and update test patches accordingly

+++ grader/utils/virtual_environment.py
@@ -22,9 +22,10 @@ class VirtualEnvironment:
 
     is_initialized = False
 
-    def __init__(self, project_path: str):
+    def __init__(self, project_path: str, keep_venv: bool = False):
         self._project_path = project_path
         self._venv_path = os.path.join(project_path, const.VENV_NAME)
+        self.__keep_venv = keep_venv
 
     def __enter__(self):
         self.setup()
@@ -81,7 +82,8 @@ class VirtualEnvironment:
         """
         Delete the virtual environment.
         """
-        shutil.rmtree(self._venv_path)
+        if not self.__keep_venv:
+            shutil.rmtree(self._venv_path)
 
     @staticmethod
     def __install_requirements(venv_path: str, requirements_path: str):
